Make compatible with V8
2 small changes to make it compatible with Umbraco V8

In 8 tabs are not stored on the current editorState but on the variants property. So updated the code to get them from the active variant when there is a variants property on the editorState.

jQuery load is deprecated in the version used in V8. So updated that as well.

I did not test my changes on V7.